agentplane egress: the watch timeout reaches the API server as an integer
The API server parses timeoutSeconds with strconv.ParseInt, so the float
default reached it as "300.0" and every watch was refused with 400. The
informer fell back to relisting on its retry, which is why the index
still filled and the proxy still decided correctly, but no change
propagated between relists and the log filled with the rejection every
30 seconds.

Observed on staging:

  informer WARNING Retrying <unknown> in 30 seconds as it raised
  ApiException: (400) Reason: BadRequest:
  strconv.ParseInt: parsing "300.0": invalid syntax

The fake API server parsed the same field as a float and so accepted
what the real one rejects; it now answers 400 on a non-integer, the way
the real one does.

+++ x/agentplane/egress/informer.py
@@ -116,7 +116,7 @@ class Informer:
         core_v1: CoreV1Api,
         namespace: str,
         credentials_namespace: str,
-        resync_seconds: float,
+        resync_seconds: int,
         clock: Callable[[], datetime] = lambda: datetime.now(UTC),
     ) -> None:
         self._index = index
@@ -197,6 +197,8 @@ class Informer:
         await self._changed(kind)
         watcher = k8s_watch.Watch()
         try:
+            # Integer seconds: the API server parses timeoutSeconds with strconv.ParseInt, so a float
+            # reaches it as "300.0" and every watch is refused with 400.
             async for event in watcher.stream(
                 kind.list, *kind.args, resource_version=version, timeout_seconds=self._resync_seconds
             ):
